requirements.txt
-i https://pypi.org/simple
cramjam==2.11.0; python_version >= '3.8'
et-xmlfile==2.0.0; python_version >= '3.8'
faker==37.8.0; python_version >= '3.9'
fastparquet==2024.11.0; python_version >= '3.9'
fsspec==2025.9.0; python_version >= '3.9'
numpy==2.3.3; python_version >= '3.11'
openpyxl==3.1.5; python_version >= '3.8'
packaging==25.0; python_version >= '3.8'
pandas==2.3.3; python_version >= '3.9'
psutil==7.1.0; python_version >= '3.6'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.2
pyyaml==6.0.3; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
tzdata==2025.2; python_version >= '2'
